Fix #4788: what's new menu item (#5422)
* Move PageActionMenu and AppMenu into their own files
Currently they are mixed in the PhotonActionMenuProtocol file.
Ultimately, that file should be emptied of non-utility extensions.

* Remove unused appversion variable

* Rename hideimagesBadge to appMenuBadge

* App menu badge depends on showing whats new

* add menu item for whats new

* Always show the getting started menu item, but only badge it when first upgraded.

* add telemetry

* UX change: move badge back to upper right (not lower right)
