+++ Client/Frontend/Widgets/PhotonActionSheet/PhotonActionSheetProtocol.swift
@@ -38,247 +38,7 @@ extension PhotonActionSheetProtocol {
         viewController.present(sheet, animated: true, completion: nil)
     }
 
-    //Returns a list of actions which is used to build a menu
-    //OpenURL is a closure that can open a given URL in some view controller. It is up to the class using the menu to know how to open it
-    func getLibraryActions(vcDelegate: PageOptionsVC) -> [PhotonActionSheetItem] {
-        guard let tab = self.tabManager.selectedTab else { return [] }
-
-        let openLibrary = PhotonActionSheetItem(title: Strings.AppMenuLibraryTitleString, iconString: "menu-library") { _, _ in
-            let bvc = vcDelegate as? BrowserViewController
-            bvc?.showLibrary()
-        }
-
-        let openHomePage = PhotonActionSheetItem(title: Strings.AppMenuOpenHomePageTitleString, iconString: "menu-Home") { _, _ in
-            let page = NewTabAccessors.getHomePage(self.profile.prefs)
-            if page == .homePage, let homePageURL = HomeButtonHomePageAccessors.getHomePage(self.profile.prefs) {
-                tab.loadRequest(PrivilegedRequest(url: homePageURL) as URLRequest)
-            } else if let homePanelURL = page.url {
-                tab.loadRequest(PrivilegedRequest(url: homePanelURL) as URLRequest)
-            }
-        }
-
-        return [openHomePage, openLibrary]
-    }
-
-    /*
-     Returns a list of actions which is used to build the general browser menu
-     These items repersent global options that are presented in the menu
-     TODO: These icons should all have the icons and use Strings.swift
-     */
-
     typealias PageOptionsVC = QRCodeViewControllerDelegate & SettingsDelegate & PresentingModalViewControllerDelegate & UIViewController
-
-    func getOtherPanelActions(vcDelegate: PageOptionsVC) -> [PhotonActionSheetItem] {
-        var items: [PhotonActionSheetItem] = []
-
-        let noImageEnabled = NoImageModeHelper.isActivated(profile.prefs)
-        let noImageMode = PhotonActionSheetItem(title: Strings.AppMenuNoImageMode, iconString: "menu-NoImageMode", isEnabled: noImageEnabled, accessory: .Switch, badgeIconNamed: "menuBadge") { action,_ in
-            NoImageModeHelper.toggle(isEnabled: action.isEnabled, profile: self.profile, tabManager: self.tabManager)
-        }
-
-        items.append(noImageMode)
-
-        let nightModeEnabled = NightModeHelper.isActivated(profile.prefs)
-        let nightMode = PhotonActionSheetItem(title: Strings.AppMenuNightMode, iconString: "menu-NightMode", isEnabled: nightModeEnabled, accessory: .Switch) { _, _ in
-            NightModeHelper.toggle(self.profile.prefs, tabManager: self.tabManager)
-            // If we've enabled night mode and the theme is normal, enable dark theme
-            if NightModeHelper.isActivated(self.profile.prefs), ThemeManager.instance.currentName == .normal {
-                ThemeManager.instance.current = DarkTheme()
-                NightModeHelper.setEnabledDarkTheme(self.profile.prefs, darkTheme: true)
-            }
-            // If we've disabled night mode and dark theme was activated by it then disable dark theme
-            if !NightModeHelper.isActivated(self.profile.prefs), NightModeHelper.hasEnabledDarkTheme(self.profile.prefs), ThemeManager.instance.currentName == .dark {
-                ThemeManager.instance.current = NormalTheme()
-                NightModeHelper.setEnabledDarkTheme(self.profile.prefs, darkTheme: false)
-            }
-        }
-        items.append(nightMode)
-
-        let openSettings = PhotonActionSheetItem(title: Strings.AppMenuSettingsTitleString, iconString: "menu-Settings") { _, _ in
-            let settingsTableViewController = AppSettingsTableViewController()
-            settingsTableViewController.profile = self.profile
-            settingsTableViewController.tabManager = self.tabManager
-            settingsTableViewController.settingsDelegate = vcDelegate
-
-            let controller = ThemedNavigationController(rootViewController: settingsTableViewController)
-            controller.presentingModalViewControllerDelegate = vcDelegate
-
-            // Wait to present VC in an async dispatch queue to prevent a case where dismissal
-            // of this popover on iPad seems to block the presentation of the modal VC.
-            DispatchQueue.main.async {
-                vcDelegate.present(controller, animated: true, completion: nil)
-            }
-        }
-        items.append(openSettings)
-
-        return items
-    }
-
-    fileprivate func shareFileURL(url: URL, file: URL?, buttonView: UIView, presentableVC: PresentableVC) {
-        let helper = ShareExtensionHelper(url: url, file: file, tab: tabManager.selectedTab)
-        let controller = helper.createActivityViewController { completed, activityType in
-            print("Shared downloaded file: \(completed)")
-        }
-
-        if let popoverPresentationController = controller.popoverPresentationController {
-            popoverPresentationController.sourceView = buttonView
-            popoverPresentationController.sourceRect = buttonView.bounds
-            popoverPresentationController.permittedArrowDirections = .up
-        }
-
-        presentableVC.present(controller, animated: true, completion: nil)
-    }
-
-    func getTabActions(tab: Tab, buttonView: UIView,
-                       presentShareMenu: @escaping (URL, Tab, UIView, UIPopoverArrowDirection) -> Void,
-                       findInPage:  @escaping () -> Void,
-                       presentableVC: PresentableVC,
-                       isBookmarked: Bool,
-                       isPinned: Bool,
-                       success: @escaping (String) -> Void) -> Array<[PhotonActionSheetItem]> {
-        if tab.url?.isFileURL ?? false {
-            let shareFile = PhotonActionSheetItem(title: Strings.AppMenuSharePageTitleString, iconString: "action_share") { _, _ in
-                guard let url = tab.url else { return }
-
-                self.shareFileURL(url: url, file: nil, buttonView: buttonView, presentableVC: presentableVC)
-            }
-
-            return [[shareFile]]
-        }
-
-        let toggleActionTitle = tab.desktopSite ? Strings.AppMenuViewMobileSiteTitleString : Strings.AppMenuViewDesktopSiteTitleString
-        let toggleDesktopSite = PhotonActionSheetItem(title: toggleActionTitle, iconString: "menu-RequestDesktopSite", isEnabled: tab.desktopSite, badgeIconNamed: "menuBadge") { _, _ in
-            if let url = tab.url {
-                tab.toggleDesktopSite()
-                Tab.DesktopSites.updateDomainList(forUrl: url, isDesktopSite: tab.desktopSite, isPrivate: tab.isPrivate)
-            }
-        }
-
-        let addReadingList = PhotonActionSheetItem(title: Strings.AppMenuAddToReadingListTitleString, iconString: "addToReadingList") { _, _ in
-            guard let url = tab.url?.displayURL else { return }
-
-            self.profile.readingList.createRecordWithURL(url.absoluteString, title: tab.title ?? "", addedBy: UIDevice.current.name)
-            UnifiedTelemetry.recordEvent(category: .action, method: .add, object: .readingListItem, value: .pageActionMenu)
-            success(Strings.AppMenuAddToReadingListConfirmMessage)
-        }
-
-        let bookmarkPage = PhotonActionSheetItem(title: Strings.AppMenuAddBookmarkTitleString, iconString: "menu-Bookmark") { _, _ in
-            guard let url = tab.canonicalURL?.displayURL,
-                let bvc = presentableVC as? BrowserViewController else {
-                return
-            }
-            bvc.addBookmark(url: url.absoluteString, title: tab.title, favicon: tab.displayFavicon)
-            UnifiedTelemetry.recordEvent(category: .action, method: .add, object: .bookmark, value: .pageActionMenu)
-            success(Strings.AppMenuAddBookmarkConfirmMessage)
-        }
-
-        let removeBookmark = PhotonActionSheetItem(title: Strings.AppMenuRemoveBookmarkTitleString, iconString: "menu-Bookmark-Remove") { _, _ in
-            guard let url = tab.url?.displayURL else { return }
-
-            self.profile.places.deleteBookmarksWithURL(url: url.absoluteString).uponQueue(.main) { result in
-                if result.isSuccess {
-                    success(Strings.AppMenuRemoveBookmarkConfirmMessage)
-                }
-            }
-
-            UnifiedTelemetry.recordEvent(category: .action, method: .delete, object: .bookmark, value: .pageActionMenu)
-        }
-
-        let pinToTopSites = PhotonActionSheetItem(title: Strings.PinTopsiteActionTitle, iconString: "action_pin") { _, _ in
-            guard let url = tab.url?.displayURL, let sql = self.profile.history as? SQLiteHistory else { return }
-
-            sql.getSites(forURLs: [url.absoluteString]).bind { val -> Success in
-                guard let site = val.successValue?.asArray().first?.flatMap({ $0 }) else {
-                    return succeed()
-                }
-
-                return self.profile.history.addPinnedTopSite(site)
-            }.uponQueue(.main) { _ in }
-        }
-
-        let removeTopSitesPin = PhotonActionSheetItem(title: Strings.RemovePinTopsiteActionTitle, iconString: "action_unpin") { _, _ in
-            guard let url = tab.url?.displayURL, let sql = self.profile.history as? SQLiteHistory else { return }
-
-            sql.getSites(forURLs: [url.absoluteString]).bind { val -> Success in
-                guard let site = val.successValue?.asArray().first?.flatMap({ $0 }) else {
-                    return succeed()
-                }
-
-                return self.profile.history.removeFromPinnedTopSites(site)
-            }.uponQueue(.main) { _ in }
-        }
-
-        let sendToDevice = PhotonActionSheetItem(title: Strings.SendToDeviceTitle, iconString: "menu-Send-to-Device") { _, _ in
-            guard let bvc = presentableVC as? PresentableVC & InstructionsViewControllerDelegate & DevicePickerViewControllerDelegate else { return }
-            if !self.profile.hasAccount() {
-                let instructionsViewController = InstructionsViewController()
-                instructionsViewController.delegate = bvc
-                let navigationController = UINavigationController(rootViewController: instructionsViewController)
-                navigationController.modalPresentationStyle = .formSheet
-                bvc.present(navigationController, animated: true, completion: nil)
-                return
-            }
-
-            let devicePickerViewController = DevicePickerViewController()
-            devicePickerViewController.pickerDelegate = bvc
-            devicePickerViewController.profile = self.profile
-            devicePickerViewController.profileNeedsShutdown = false
-            let navigationController = UINavigationController(rootViewController: devicePickerViewController)
-            navigationController.modalPresentationStyle = .formSheet
-            bvc.present(navigationController, animated: true, completion: nil)
-        }
-
-        let sharePage = PhotonActionSheetItem(title: Strings.AppMenuSharePageTitleString, iconString: "action_share") { _, _ in
-            guard let url = tab.canonicalURL?.displayURL else { return }
-
-            if let temporaryDocument = tab.temporaryDocument {
-                temporaryDocument.getURL().uponQueue(.main, block: { tempDocURL in
-                    // If we successfully got a temp file URL, share it like a downloaded file,
-                    // otherwise present the ordinary share menu for the web URL.
-                    if tempDocURL.isFileURL {
-                        self.shareFileURL(url: url, file: tempDocURL, buttonView: buttonView, presentableVC: presentableVC)
-                    } else {
-                        presentShareMenu(url, tab, buttonView, .up)
-                    }
-                })
-            } else {
-                presentShareMenu(url, tab, buttonView, .up)
-            }
-        }
-
-        let copyURL = PhotonActionSheetItem(title: Strings.AppMenuCopyURLTitleString, iconString: "menu-Copy-Link") { _, _ in
-            if let url = tab.canonicalURL?.displayURL {
-                UIPasteboard.general.url = url
-                success(Strings.AppMenuCopyURLConfirmMessage)
-            }
-        }
-
-        var mainActions = [sharePage]
-
-        // Disable bookmarking and reading list if the URL is too long.
-        if !tab.urlIsTooLong {
-            mainActions.append(isBookmarked ? removeBookmark : bookmarkPage)
-
-            if tab.readerModeAvailableOrActive {
-                mainActions.append(addReadingList)
-            }
-        }
-
-        mainActions.append(contentsOf: [sendToDevice, copyURL])
-
-        let pinAction = (isPinned ? removeTopSitesPin : pinToTopSites)
-        var commonActions = [toggleDesktopSite, pinAction]
-
-        // Disable find in page if document is pdf.
-        if tab.mimeType != MIMEType.PDF {
-            let findInPageAction = PhotonActionSheetItem(title: Strings.AppMenuFindInPageTitleString, iconString: "menu-FindInPage") { _, _ in
-                findInPage()
-            }
-            commonActions.insert(findInPageAction, at: 0)
-        }
-
-        return [mainActions, commonActions]
-    }
 
     func fetchBookmarkStatus(for url: String) -> Deferred<Maybe<Bool>> {
         return profile.places.isBookmarked(url: url)
@@ -341,48 +101,4 @@ extension PhotonActionSheetProtocol {
         }
     }
 
-    func syncMenuButton(showFxA: @escaping (_ params: FxALaunchParams?) -> Void) -> PhotonActionSheetItem? {
-        profile.getAccount()?.updateProfile()
-        let account = profile.getAccount()
-
-        func title() -> String? {
-            guard let status = account?.actionNeeded else { return Strings.FxASignInToSync }
-            switch status {
-            case .none:
-                return account?.fxaProfile?.displayName ?? account?.fxaProfile?.email
-            case .needsVerification:
-                return Strings.FxAAccountVerifyEmail
-            case .needsPassword:
-                return Strings.FxAAccountVerifyPassword
-            case .needsUpgrade:
-                return Strings.FxAAccountUpgradeFirefox
-            }
-        }
-
-        func imageName() -> String? {
-            guard let status = account?.actionNeeded else { return "menu-sync" }
-            switch status {
-            case .none:
-                return "placeholder-avatar"
-            case .needsVerification, .needsPassword, .needsUpgrade:
-                return "menu-warning"
-            }
-        }
-
-        let action: ((PhotonActionSheetItem, UITableViewCell) -> Void) = { action,_ in
-            let fxaParams = FxALaunchParams(query: ["entrypoint": "browsermenu"])
-            showFxA(fxaParams)
-        }
-
-        guard let title = title(), let iconString = imageName() else { return nil }
-        guard let actionNeeded = account?.actionNeeded else {
-            let signInOption = PhotonActionSheetItem(title: title, iconString: iconString, handler: action)
-            return signInOption
-        }
-
-        let iconURL = (actionNeeded == .none) ? account?.fxaProfile?.avatar.url : nil
-        let iconType: PhotonActionSheetIconType = (actionNeeded == .none) ? .URL : .Image
-        let syncOption = PhotonActionSheetItem(title: title, iconString: iconString, iconURL: iconURL, iconType: iconType, accessory: .Sync, handler: action)
-        return syncOption
-    }
 }
